CELE-109 Fixes screen flash with 3D viewer
The problem of flash came from the fact that, depending on the dataset,
the data couldn't be fetch. When that was the case, none of the scene
was rendered, even if only one file couldn't be loaded. This commit
changes the way the 3D images are loaded, and shows an error message to
inform the user if one of the 3D images cannot be displayed (the other
images for the requested neurons are displayed if they exists in the
bucket).