CELE-144 Fix issue with react-color picker moving in the synapse tree
Solution have been implemented by Claude, the original problem came from
the fact that as we have a dynamic recomputation of the color in the
tree, changing the color was triggering a redraw, destroying the
anchorElement for the color picker. The solution Claude implemented
considers a virtual anchorElement which is created when the color picker
is opened and cleaned when the color picker is closed. This virtual
anchorElement provides a way to have an element that is not "moving"
even where there is color recomputation and re-rendering